feat: implements dry-run option for commands

+++ src/forge/gitea.rs
@@ -241,6 +241,13 @@ impl Gitea {
     async fn delete_branch_if_exists(&self, branch: &str) -> Result<()> {
         let url = self.base_url.join(&format!("branches/{branch}"))?;
         let request = self.client.delete(url).build()?;
+        if self.config.dry_run {
+            warn!(
+                "dry_run: would execute request: method: delete, url: {:?}",
+                request
+            );
+            return Ok(());
+        }
         let resp = self.client.execute(request).await?;
         info!("delete branch resp --> {:#?}", resp);
         Ok(())
@@ -267,6 +274,10 @@ impl Gitea {
     }
 
     async fn create_label(&self, label_name: String) -> Result<Label> {
+        if self.config.dry_run {
+            warn!("dry_run: would create label: {label_name}");
+            return Ok(Label::default());
+        }
         let labels_url = self.base_url.join("labels")?;
         let request = self
             .client
@@ -448,6 +459,10 @@ impl Forge for Gitea {
         &self,
         req: CreateBranchRequest,
     ) -> Result<Commit> {
+        if self.config.dry_run {
+            warn!("dry_run: would create release branch: req: {:#?}", req);
+            return Ok(Commit { sha: "fff".into() });
+        }
         // TODO: Once below issue is resolved we can delete this call
         // https://github.com/go-gitea/gitea/issues/35538
         self.delete_branch_if_exists(&req.branch).await?;
@@ -498,6 +513,10 @@ impl Forge for Gitea {
     }
 
     async fn tag_commit(&self, tag_name: &str, sha: &str) -> Result<()> {
+        if self.config.dry_run {
+            warn!("dry_run: would create tag: {tag_name}",);
+            return Ok(());
+        }
         let tag_url = self.base_url.join("tags")?;
         let body = serde_json::json!({
           "tag_name": tag_name,
@@ -635,6 +654,15 @@ impl Forge for Gitea {
     }
 
     async fn create_pr(&self, req: CreatePrRequest) -> Result<PullRequest> {
+        if self.config.dry_run {
+            warn!("dry_run: would create release PR: req: {:#?}", req);
+            return Ok(PullRequest {
+                number: 0,
+                sha: "fff".into(),
+                body: req.body,
+            });
+        }
+
         let data = CreatePull {
             title: req.title,
             body: req.body,
@@ -655,6 +683,10 @@ impl Forge for Gitea {
     }
 
     async fn update_pr(&self, req: UpdatePrRequest) -> Result<()> {
+        if self.config.dry_run {
+            warn!("dry_run: would update release PR: req: {:#?}", req);
+            return Ok(());
+        }
         let data = UpdatePullBody {
             title: req.title,
             body: req.body,
@@ -669,6 +701,10 @@ impl Forge for Gitea {
     }
 
     async fn replace_pr_labels(&self, req: PrLabelsRequest) -> Result<()> {
+        if self.config.dry_run {
+            warn!("dry_run: would replace release PR labels: req: {:#?}", req);
+            return Ok(());
+        }
         let all_labels = self.get_all_labels().await?;
 
         let mut labels = vec![];
@@ -701,6 +737,12 @@ impl Forge for Gitea {
         sha: &str,
         notes: &str,
     ) -> Result<()> {
+        if self.config.dry_run {
+            warn!(
+                "dry_run: would create release: tag: {tag}, sha: {sha}, notes: {notes}"
+            );
+            return Ok(());
+        }
         let data = CreateRelease {
             tag_name: tag.to_string(),
             target_commitish: sha.to_string(),
